requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
Pillow==8.3.1
PyJWT==v1.7.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
